Extract code into DiagnosisKeyBatchProcessor
